Logic tab order updates and hook fixes
=== General

Updated the logical tab order for the new react hooks to be more left to
right instead of jumping all around.

Fixed the capitalization for the useState hook

Updated the useContext hook to include a tab to remove the entire
`const context` part

Updated the useCallback and useMemo to be able to quickly add arguments
or return the value immediately (optional surrounding braces {})

=== Typescript
Update `sdp` to no longer be typed. There is now a `sdpt` to create a typed
static default props version

Updated `cdp` to no longer be typed. There is now a `cdpt` to create a typed
component default props version.